+++ module/helpers/templates.mjs
@@ -4,7 +4,7 @@
  * @return {Promise}
  */
 export const preloadHandlebarsTemplates = async function () {
-  return loadTemplates([
+  return foundry.applications.handlebars.loadTemplates([
     // Actor partials.
     "systems/knave2e/templates/actor/parts/actor-character-items.hbs",
   ]);