+++ module/knave2e.mjs
@@ -318,10 +318,10 @@ Hooks.once('init', function () {
     CONFIG.ChatMessage.documentClass = Knave2eChatMessage;
 
     // Register sheet application classes
-    Actors.unregisterSheet('core', ActorSheet);
-    Actors.registerSheet('knave2e', Knave2eActorSheet, { makeDefault: true });
-    Items.unregisterSheet('core', ItemSheet);
-    Items.registerSheet('knave2e', Knave2eItemSheet, { makeDefault: true });
+    foundry.documents.collections.Actors.unregisterSheet('core', foundry.appv1.sheets.ActorSheet);
+    foundry.documents.collections.Actors.registerSheet('knave2e', Knave2eActorSheet, { makeDefault: true });
+    foundry.documents.collections.Items.unregisterSheet('core', foundry.appv1.sheets.ItemSheet);
+    foundry.documents.collections.Items.registerSheet('knave2e', Knave2eItemSheet, { makeDefault: true });
 
     // Preload Handlebars templates.
     return preloadHandlebarsTemplates();